Moves pagination logic from controller to new module
